requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
alembic>=1.14.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic[email]>=2.10.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.0,<2.0.0
python-multipart>=0.0.18,<1.0.0
bcrypt>=4.2.0,<5.0.0
slowapi>=0.1.9,<1.0.0
